+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Add POSIX-compliant low-level driver
 - Prohibit transmission of too long UDP packets (default), with conditional compilation
 - Split CMakeLists.txt files between library and executable
+- Move `esp_set_server` function to separate file
+- Run `AT+GMR` command just after reset/restore to determine ESP device connected
+- Min required AT binaries are now `2.3.0` for `ESP32/ESP32C3` and `2.2.1` for `ESP8266`
 
 ## v1.1.1-dev